Significantly sped up the neighbour list
Abandoned trying to use a jax based approach as it is too difficult with
the need to avoid dynamically sized arrays.  Using a simple but fast
numpy based approach instead.  It's about 2-5 times faster than the ASE
neighbour list, although this probably comes at the cost of a higher
memory requirement.