Changed the way Sequential works for partials
This affects the way that, say, Grads are used in a sequence of layers.
Before it was placed at the end of the sequence and re-ordered by
Sequantial to be placed at the start.  This doesn't allow enough
flexibility if someone, say, wants to put a partial half way through a
Sequence of layers.  Now, it will remain in the position it is in, and
everything after will be bundled up and passed to the partial as a
Module.