Moved diff() out of experimental
The diff() function has proved to be super useful so is being brought of
out experimenta.
Also updated NodewiseEmbedding to be able to use broadcasting meaning
that, e.g., one could use a global.key where the corresponding value is
does now have a shape that corresponds to the number of graphs (which it
should), in which case this will be broadcasted 'up'.